Added initial components option to Bootstrap init
Added contains method to MapServiceLocator
to be able to check if a class has already been resolved from the initialComponents.

String argumented Bootstrap init method will now also check if the file resides
in the resources or throw an exception.

Added Spec2 and logback to the http test classpath. Added the simple Bootstrap tests.

Removed credentials key from the build. It is expected to have it declared in the sbt Global.