Added initialComponents argument to Bootstrap init
String argumented Bootstrap init will now try to load
properties from resources if they are not present in the system,
or throw an RuntimeException if they are not present there also.

Added HttpHeaderProvider as an possible initialComponent,
with default to SettingsHeaderProvider with reads from properties.
HttpClient now depends on HttpHeaderProvider to provide it with
authorization headers.

Removed ProjectSettings, java.util.Properties are now a part of the
container.

Added simple specification for Bootstrap and HttpHeaderProvider.