JsonSerialization Tests.
An instance of database is expected to exist.
Under user, with password and named 'scalaclienttest', on localhost:5432.

On sbt load dsl will be compiled from resources of http project, if it's not there already.
Jar will be placed in the http's test classpath.

On test call.
Local instance of revenj will be started.
It will connect to above-mentioned database.
On test complete, revenj will be shutdown, if it started it.

Process will not check if the 8999 port is already used.
Nore will it stop if its instance hasn't started successfuly.
It will then use a service currently running at the port 8999 to complete tests.